chore(#222): fix some strings in portuguese translation
Some words have more than one meaning (`right` is the opposite of
`left`, but it's also the opposite of `wrong`). Also, in English in some
cases the same word is a verb or a noun. The translation needs to use
the correct form.

Apparently the initial translation used some tool to generate the
translation. This tool used the wrong meaning in some places, resulting
in strange sentences.

This PR fixes a few translation errors.